:hammer:Update time to long

+++ openjob-server/openjob-server-repository/src/main/java/io/openjob/server/repository/dao/impl/JobDAOImpl.java
@@ -30,7 +30,7 @@ public class JobDAOImpl implements JobDAO {
     }
 
     @Override
-    public List<Job> listScheduledJobs(List<Long> slotIds, Integer time) {
+    public List<Job> listScheduledJobs(List<Long> slotIds, Long time) {
         List<String> notTypes = Arrays.asList(TimeExpressionTypeEnum.NONE.getType(), TimeExpressionTypeEnum.SECOND_DELAY.getType());
         return jobRepository.findBySlotsIdInAndStatusAndTimeExpressionTypeNotInAndNextExecuteTimeLessThanEqual(slotIds, JobStatusEnum.RUNNING.getStatus(), notTypes, time);
     }
