+++ openjob-server/openjob-server-repository/src/main/java/io/openjob/server/repository/repository/DelayInstanceRepository.java
@@ -28,7 +28,7 @@ public interface DelayInstanceRepository extends JpaRepository<DelayInstance, Lo
     @Transactional(rollbackFor = Exception.class)
     @Modifying
     @Query(value = "update DelayInstance as d set d.status=?2, d.updateTime=?3 where d.id in(?1)")
-    Integer batchUpdateStatus(List<Long> ids, Integer status, Integer updateTime);
+    Integer batchUpdateStatus(List<Long> ids, Integer status, Long updateTime);
 
     /**
      * Delete by task id.
