+++ openjob-server/openjob-server-repository/src/main/java/io/openjob/server/repository/repository/ServerReportsRepository.java
@@ -17,5 +17,5 @@ public interface ServerReportsRepository extends JpaRepository<ServerReports, Lo
      * @param status     status
      * @return Long
      */
-    Long countByCreateTimeGreaterThanEqualAndServerIdAndStatus(Integer createTime, Long serverId, Integer status);
+    Long countByCreateTimeGreaterThanEqualAndServerIdAndStatus(Long createTime, Long serverId, Integer status);
 }
